Refactor: userService 코드 정리 #4
