+++ src/main/java/com/coastee/server/login/facade/LoginFacade.java
@@ -1,4 +1,4 @@
-package com.coastee.server.login.service;
+package com.coastee.server.login.facade;
 
 import com.coastee.server.auth.domain.Accessor;
 import com.coastee.server.global.apipayload.exception.handler.InvalidJwtException;
@@ -7,28 +7,26 @@ import com.coastee.server.login.domain.OAuthLoginParams;
 import com.coastee.server.login.domain.OAuthUserInfo;
 import com.coastee.server.login.infrastructure.JwtProvider;
 import com.coastee.server.login.infrastructure.loginparams.LinkedInLoginParams;
+import com.coastee.server.login.service.OAuthInfoService;
 import com.coastee.server.user.domain.User;
-import com.coastee.server.user.domain.repository.UserRepository;
 import com.coastee.server.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import static com.coastee.server.global.apipayload.code.status.ErrorStatus.FAIL_VALIDATE_TOKEN;
-import static com.coastee.server.global.apipayload.code.status.ErrorStatus.INVALID_REFRESH_TOKEN;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class LoginService {
-    private final RequestOAuthInfoService requestOAuthInfoService;
+public class LoginFacade {
+    private final OAuthInfoService oAuthInfoService;
     private final UserService userService;
-    private final UserRepository userRepository;
     private final JwtProvider jwtProvider;
 
     public AuthTokens login(final OAuthLoginParams params) {
-        OAuthUserInfo userInfo = requestOAuthInfoService.request(params);
-        User user = findOrCreateUser(userInfo);
+        OAuthUserInfo userInfo = oAuthInfoService.request(params);
+        User user = userService.findOrCreateUser(userInfo);
         AuthTokens tokens = jwtProvider.createTokens(user.getId().toString());
         user.updateRefreshToken(tokens.getRefreshToken());
         return tokens;
@@ -39,29 +37,13 @@ public class LoginService {
             final LinkedInLoginParams params
     ) {
         User user = userService.findById(accessor.getUserId());
-        OAuthUserInfo userInfo = requestOAuthInfoService.request(params);
+        OAuthUserInfo userInfo = oAuthInfoService.request(params);
         user.verify(userInfo.getSocialId());
-    }
-
-    private User findOrCreateUser(final OAuthUserInfo userInfo) {
-        return userRepository.findBySocialId(userInfo.getSocialId())
-                .orElseGet(() -> newUser(userInfo));
-    }
-
-    private User newUser(final OAuthUserInfo userInfo) {
-        User user = User.of()
-                .nickname(userInfo.getName())
-                .email(userInfo.getEmail())
-                .socialType(userInfo.getSocialType())
-                .socialId(userInfo.getSocialId())
-                .build();
-        return userRepository.save(user);
     }
 
     public String renewalToken(final String accessToken, final String refreshToken) {
         if (jwtProvider.isValidRefreshAndExpiredAccess(refreshToken, accessToken)) {
-            User user = userRepository.findByRefreshToken(refreshToken)
-                    .orElseThrow(() -> new InvalidJwtException(INVALID_REFRESH_TOKEN));
+            User user = userService.findByRefreshToken(refreshToken);
             return jwtProvider.createAccessToken(user.getId().toString());
         }
         if (jwtProvider.isValidRefreshAndValidAccess(refreshToken, accessToken)) {
